Fix description on how a docker container is started in k8s
In k8s we set `securityContext.allowPrivilegeEscalation=false` for the container.

Due to this, a user with `sudo` rights would start fine locally with `docker run` but fail to start in a binderhub k8s pod with allowPrivilegeEscalation=False. To represent this better, we can add `--security-opt=no-new-privileges:true` to the `docker run` command. Like that I would get the error: `standard_init_linux.go:211: exec user process caused "operation not permitted"` instead.